Created Timer.jsx to separate the timer component out.
